# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stellar NFT Minter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MintRecordStatus = typeof MintRecordStatus[keyof typeof MintRecordStatus];


export const MintRecordStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface MintRecord {
  id: number;
  walletAddress: string;
  nftName: string;
  description: string;
  imageUrl: string;
  /**
     * JSON string of additional attributes
     * @nullable
     */
  attributes?: string | null;
  status: MintRecordStatus;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  nftId?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  contractAddress?: string | null;
  network?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface MintInput {
  /** @minLength 1 */
  walletAddress: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  nftName: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  description: string;
  /** @minLength 1 */
  imageUrl: string;
  /** JSON string of additional attributes */
  attributes?: string;
  contractAddress?: string;
  network: string;
}

export type MintUpdateStatus = typeof MintUpdateStatus[keyof typeof MintUpdateStatus];


export const MintUpdateStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface MintUpdate {
  status?: MintUpdateStatus;
  txHash?: string;
  nftId?: string;
  errorMessage?: string;
  contractAddress?: string;
}

export interface MintStats {
  totalMints: number;
  successfulMints: number;
  failedMints: number;
  pendingMints: number;
  uniqueWallets: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListMintsParams = {
/**
 * Filter by wallet address
 */
walletAddress?: string;
/**
 * Maximum number of results
 */
limit?: number;
/**
 * Offset for pagination
 */
offset?: number;
};