# lib/api-zod/src/generated/types/mintInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stellar NFT Minter API
 * OpenAPI spec version: 0.1.0
 */

export interface MintInput {
  /** @minLength 1 */
  walletAddress: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  nftName: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  description: string;
  /** @minLength 1 */
  imageUrl: string;
  /** JSON string of additional attributes */
  attributes?: string;
  contractAddress?: string;
  network: string;
}