# lib/api-zod/src/generated/types/mintRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stellar NFT Minter API
 * OpenAPI spec version: 0.1.0
 */
import type { MintRecordStatus } from './mintRecordStatus';

export interface MintRecord {
  id: number;
  walletAddress: string;
  nftName: string;
  description: string;
  imageUrl: string;
  /**
     * JSON string of additional attributes
     * @nullable
     */
  attributes?: string | null;
  status: MintRecordStatus;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  nftId?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  contractAddress?: string | null;
  network?: string;
  createdAt: string;
  updatedAt?: string;
}