# lib/api-zod/src/generated/types/mintStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stellar NFT Minter API
 * OpenAPI spec version: 0.1.0
 */

export interface MintStats {
  totalMints: number;
  successfulMints: number;
  failedMints: number;
  pendingMints: number;
  uniqueWallets: number;
}