# lib/api-zod/src/generated/types/mintUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stellar NFT Minter API
 * OpenAPI spec version: 0.1.0
 */
import type { MintUpdateStatus } from './mintUpdateStatus';

export interface MintUpdate {
  status?: MintUpdateStatus;
  txHash?: string;
  nftId?: string;
  errorMessage?: string;
  contractAddress?: string;
}